jobs/requirements.txt
ansible==2.8.0
apipkg==1.5
appdirs==1.4.3
argh==0.26.2
asn1crypto==0.24.0
async-generator==1.10
asyncio-extras==1.3.2
atomicwrites==1.3.0
attrs==18.1.0
autopep8==1.4.4
awscli==1.16.167
backcall==0.1.0
bcrypt==3.1.6
black==19.3b0
boto3==1.9.157
botocore==1.12.157
canonicalwebteam.snapstoreapi==0.4.4
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.6.1
decorator==4.4.0
defusedxml==0.6.0
distro==1.4.0
docopt==0.6.2
docutils==0.14
easywatch==0.0.5
entrypoints==0.3
execnet==1.6.0
fasteners==0.15
flake8==3.7.7
httplib2==0.12.3
idna==2.8
importlib-metadata==0.15
invoke==1.2.0
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.13.3
jeepney==0.4
jenkins-job-builder==2.10.0
Jinja2==2.10.1
jmespath==0.9.4
juju==0.11.7
juju-wait==2.7.0
jujubundlelib==0.5.6
keyring==19.0.2
kv==0.3
launchpadlib==1.10.6
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.2.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
more-itertools==7.0.0
multi-key-dict==2.0.3
oauthlib==3.0.1
paramiko==2.4.2
parso==0.4.0
pathtools==0.1.2
pbr==5.2.0
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycountry==18.12.8
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pymacaroons==0.13.0
PyNaCl==1.1.2
pyRFC3339==1.1
pytest==4.4.0
pytest-asyncio==0.10.0
pytest-forked==1.0.2
pytest-xdist==1.28.0
python-box==3.2.3
python-dateutil==2.8.0
python-dotenv==0.10.2
python-jenkins==1.4.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==3.13
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.0
SecretStorage==3.1.1
semver==2.8.1
sh==1.12.14
six==1.12.0
staticjinja==0.3.5
stevedore==1.30.1
testresources==2.0.1
theblues==0.5.1
toml==0.10.0
toolz==0.9.0
traitlets==4.3.2
urllib3==1.25.3
wadllib==1.3.3
watchdog==0.9.0
wcwidth==0.1.7
websockets==7.0
zipp==0.5.1
ogc==0.1.5
